Reworked damage system to rely on floats and added Health Regen support to Damageable.cs, which can be enabled on any game object that implements it by setting a non-zero regen value on the public script properties for a prefab game object.
Code changes mainly entailed editing Damageable.cs by adding a public regen value float, calling Heal() on Update() accordingly, and changing Health int types to float types everywhere except the UI text labels which cast the new floats to an int for clean text display.

+++ Assets/Scripts/Damage/DamageAbstractBehaviors.cs
@@ -38,7 +38,7 @@ public abstract class OnDamageBehavior : MonoBehaviour
             damageable.OnDamaged -= OnDamaged;
     }
 
-    protected abstract void OnDamaged(int damage);
+    protected abstract void OnDamaged(float damage);
 }
 
 [RequireComponent(typeof(Damageable))]
@@ -57,5 +57,5 @@ public abstract class OnHealthChangeBehavior : MonoBehaviour
         damageable.OnHealthChange -= OnHealthChange;
     }
 
-    protected abstract void OnHealthChange(int health);
+    protected abstract void OnHealthChange(float health);
 }
